Migrate to new BlockStorage structure
