+++ targets/FreeRTOS/NXP_MIMXRT1060_EVK/CMakeLists.txt
@@ -51,6 +51,7 @@ add_executable(
     ${NANOBOOTER_PROJECT_NAME}.elf
 
     "${CMAKE_CURRENT_SOURCE_DIR}/target_common.c"
+    "${CMAKE_CURRENT_SOURCE_DIR}/target_BlockStorage.c"
 
     ${COMMON_PROJECT_SOURCES}
     ${NANOBOOTER_PROJECT_SOURCES}
