+++ targets/FreeRTOS/NXP_MIMXRT1060_EVK/target_BlockStorage.h
@@ -1,0 +1,12 @@
+//
+// Copyright (c) 2019 The nanoFramework project contributors
+// See LICENSE file in the project root for full license information.
+//
+
+#ifndef _TARGETPAL_BLOCKSTORAGE_H_
+#define _TARGETPAL_BLOCKSTORAGE_H_ 1
+
+// this device has 1 block storage devices
+#define TARGET_BLOCKSTORAGE_COUNT   1
+
+#endif //_TARGETPAL_BLOCKSTORAGE_H_
